requirements.txt
asgiref==3.11.1
attrs==26.1.0
beautifulsoup4==4.14.3
Django==6.0.3
django-filter==25.2
djangorestframework==3.17.1
drf-spectacular==0.29.0
et_xmlfile==2.0.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lxml==6.0.2
numpy==2.4.3
ofxparse==0.21
openpyxl==3.1.5
pandas==3.0.1
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
djangorestframework-simplejwt==5.5.1
PyJWT==2.12.1
